backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=25.2
psycopg2-binary>=2.9
python-decouple>=3.8
dj-database-url>=2.1
gunicorn>=21.2
whitenoise>=6.6
Pillow>=10.1
